chore: 폴더를 파일로 변경
