--- src/entities/auth/apis/signOutUser.tsx
@@ -1,7 +1,0 @@
-import { supabase } from '@/shared';
-
-const signOutUser = async () => {
-  return await supabase.auth.signOut();
-};
-
-export default signOutUser;
