+++ src/entities/auth/index.ts
@@ -1,9 +1,2 @@
-//apis
-export { default as signUpUser } from "./apis/signUpUser";
-export { default as signInUser } from "./apis/signInUser";
-export { default as signOutUser } from "./apis/signOutUser";
-export { default as getSession } from "./apis/getSession";
-
-// models
-
-export * from "./atoms";
+export * from './atoms';
+export * from './apis';
