+++ src/features/loginForm/index.ts
@@ -4,6 +4,3 @@ export { default as LoginForm } from './ui/LoginForm';
 // models
 export { default as useLoginMutation } from './model/useLoginMutation';
 export { default as useSignUpMutation } from './model/useSignUpMutation';
-
-export * from './atoms';
-export * from './api';
