+++ src/features/profileForm/index.ts
@@ -1,7 +1,7 @@
-export { default as ProfileForm } from "./ui/ProfileForm";
+export { default as ProfileForm } from './ui/ProfileForm';
 
 // models
-export { default as useUserNameMutation } from "./model/useUserNameMutation";
+export { default as useUserNameMutation } from './model/useUserNameMutation';
 
 // apis
-export { default as updateUserName } from "./api/updateUserName";
+export * from './apis';
